perf: defer dashboard rentals graph
